packet: go tickless

+++ src/packet.hh
@@ -6,10 +6,10 @@ class Packet
 public:
   unsigned int src;
   unsigned int flow_id;
-  unsigned int tick_sent, tick_received;
+  double tick_sent, tick_received;
 
   Packet( const unsigned int & s_src,
-	  const unsigned int & s_flow_id, const unsigned int & s_tick_sent )
+	  const unsigned int & s_flow_id, const double & s_tick_sent )
     : src( s_src ),
       flow_id( s_flow_id ), tick_sent( s_tick_sent ),
       tick_received( -1 )
